Add ExpirationToken

+++ OrderManagement/Migrations/Order/20250202125626_Init.cs
@@ -37,7 +37,7 @@ namespace OrderManagement.Migrations.Order
                     Id = table.Column<long>(type: "bigint", nullable: false),
                     CustomerId = table.Column<Guid>(type: "uniqueidentifier", nullable: false),
                     IssueDate = table.Column<DateTime>(type: "datetime2", nullable: false),
-                    Status = table.Column<byte>(type: "tinyint", nullable: false)
+                    Status = table.Column<int>(type: "int", nullable: false)
                 },
                 constraints: table =>
                 {
